Move cache seeding logic.
